Added automatic Assembly Reloading (filewatch)
