+++ Engine/src/Core/Engine.cpp
@@ -10,25 +10,28 @@
 
 namespace Pit {
 
-	InitEvent					Engine::InitEvent;
-	NetworkingUpdateEvent		Engine::NetworkingUpdateEvent;
-	PhysicUpdateEvent			Engine::PhysicUpdateEvent;
-	PreUpdateEvent				Engine::PreUpdateEvent;
-	UpdateEvent					Engine::UpdateEvent;
-	PostUpdateEvent				Engine::PostUpdateEvent;
-	RenderEvent					Engine::RenderEvent;
-	UIRenderEvent				Engine::UIRenderEvent;
-	OnWindowResizeEvent			Engine::OnWindowResizeEvent;
-	SaveConfigEvent				Engine::SaveConfigEvent;
-	ShutdownEvent				Engine::ShutdownEvent;
+	InitEvent						Engine::InitEvent;
+	NetworkingUpdateEvent			Engine::NetworkingUpdateEvent;
+	PhysicUpdateEvent				Engine::PhysicUpdateEvent;
+	PreUpdateEvent					Engine::PreUpdateEvent;
+	UpdateEvent						Engine::UpdateEvent;
+	PostUpdateEvent					Engine::PostUpdateEvent;
+	RenderEvent						Engine::RenderEvent;
+	UIRenderEvent					Engine::UIRenderEvent;
+	OnWindowResizeEvent				Engine::OnWindowResizeEvent;
+	SaveConfigEvent					Engine::SaveConfigEvent;
+	ShutdownEvent					Engine::ShutdownEvent;
 
-	EngineSettings*				Engine::s_Settings = nullptr;
+	EngineSettings*					Engine::s_Settings = nullptr;
 
-	SubmoduleManager*			Engine::s_SubmoduleManager = nullptr;
+	SubmoduleManager*				Engine::s_SubmoduleManager = nullptr;
 
-	std::ofstream				Engine::s_InstanceLockFile;
+	std::ofstream					Engine::s_InstanceLockFile;
 
-	bool						Engine::s_Quit = false;
+	Array<std::function<void()>>	Engine::s_MainThreadQueue;
+	std::mutex						Engine::s_MainThreadQueueMutex;
+
+	bool							Engine::s_Quit = false;
 
 #define CATCH_EXCEPTIONS() \
 		catch ([[maybe_unused]] const std::bad_alloc& e)	{						\
@@ -157,6 +160,8 @@ namespace Pit {
 			Time::SetDeltaTime(duration_cast<nanoseconds>(now - lastUpdate).count() * .000000001f);
 			lastUpdate = now;
 
+			ExecuteMainThreadQueue();
+
 			Input::Update();
 
 			s_SubmoduleManager->Update();
@@ -187,6 +192,18 @@ namespace Pit {
 		else
 			return false;
 	}
+
+
+	void Engine::SubmitToMainThread(std::function<void()> callback) {
+		std::scoped_lock lock(s_MainThreadQueueMutex);
+		s_MainThreadQueue.emplace_back(callback);
+	}
+
+	void Engine::ExecuteMainThreadQueue() {
+		for (auto& callback : s_MainThreadQueue) callback();
+		s_MainThreadQueue.clear();
+	}
+
 
 	AudioSubmodule* Engine::Audio()						{ if (s_SubmoduleManager) return s_SubmoduleManager->AudioSubmodule; else return nullptr; }
 	AssetManagmentSubmodule* Engine::AssetManagment()	{ if (s_SubmoduleManager) return s_SubmoduleManager->AssetManagmentSubmodule;  else return nullptr; }
