+++ Engine/src/Scripting/ScriptingSubmodule.hpp
@@ -6,89 +6,14 @@
 extern "C" {
 	typedef struct _MonoObject MonoObject;
 	typedef struct _MonoClass MonoClass;
-	typedef struct _MonoMethod MonoMethod;
 	typedef struct _MonoImage MonoImage;
-	typedef struct _MonoDomain MonoDomain;
-	typedef struct _MonoAssembly MonoAssembly;
-	typedef struct _MonoClassField MonoClassField;
 }
 
 namespace Pit {
-	enum class ScriptFieldType {
-		None = 0,
-		Float, Double,
-		Bool, Char, Byte, Short, Int, Long,
-		UByte, UShort, UInt, ULong
-	};
-
-	struct ScriptField {
-		ScriptFieldType Type;
-		std::string Name;
-
-		MonoClassField* ClassField;
-	};
-
-	// ScriptField + data storage
-	struct ScriptFieldInstance {
-		ScriptField Field;
-
-		ScriptFieldInstance() {
-			memset(m_Buffer, 0, sizeof(m_Buffer));
-		}
-
-		template<typename T>
-		T GetValue() {
-			static_assert(sizeof(T) <= 16, "Type too large!");
-			return *(T*)m_Buffer;
-		}
-
-		template<typename T>
-		void SetValue(T value) {
-			static_assert(sizeof(T) <= 16, "Type too large!");
-			std::memcpy(m_Buffer, &value, sizeof(T));
-		}
-	private:
-		uint8_t m_Buffer[16];
-
-		friend class ScriptingSubmodule;
-		friend class ScriptInstance;
-	};
-
-	using ScriptFieldMap = std::unordered_map<std::string, ScriptFieldInstance>;
-
-	class ScriptClass {
-	public:
-		ScriptClass() = default;
-		ScriptClass(const String& classNamespace, const String& className, bool isCore = false);
-
-		MonoMethod* GetMethod(const String& name, int paramCount);
-
-		MonoObject* Instantiate();
-
-		MonoClass* GetNative();
-
-	private:
-		String m_ClassNamespace, m_ClassName;
-		std::map<std::string, ScriptField> m_Fields;
-		MonoClass* m_MonoClass = nullptr;
-
-		friend class ScriptingSubmodule;
-	};
-
-	class ScriptInstance {
-	public:
-		ScriptInstance() = default;
-		ScriptInstance(MonoClass* scriptClass);
-		// TODO deletion of MonoObject ~ScriptInstance() {}
-
-		void Invoke(MonoMethod* method, void** params);
-
-		MonoClass* GetClass();
-
-	private:
-		MonoClass* m_Class = nullptr;
-		MonoObject* m_Instance = nullptr;
-	};
+	namespace Scripting {
+		class ScriptClass;
+		class ScriptInstance;
+	}
 
 	class ScriptingSubmodule : public Submodule {
 	public:
@@ -103,6 +28,7 @@ namespace Pit {
 		const bool ComponentClassExists(const String& className) const;
 
 		MonoImage* GetCoreAssemblyImage();
+		MonoImage* GetAppAssemblyImage();
 
 	private:
 		void InitMono(), ShutdownMono();
@@ -110,7 +36,7 @@ namespace Pit {
 		static MonoObject* InstantiateClass(MonoClass* monoClass);
 		void LoadAssemblyClasses();
 
-		friend class ScriptClass;
-		friend class ScriptInstance;
+		friend Scripting::ScriptClass;
+		friend Scripting::ScriptInstance;
 	};
 }
