+++ imgui.ini
@@ -30,7 +30,7 @@ Size=342,350
 Collapsed=0
 
 [Window][Profiler]
-Pos=731,432
+Pos=739,478
 Size=413,304
 Collapsed=0
 
@@ -129,7 +129,7 @@ Size=104,26
 Collapsed=0
 
 [Docking][Data]
-DockSpace       ID=0x09EF459F Window=0x9A404470 Pos=164,211 Size=1920,1056 Split=X Selected=0x13926F0B
+DockSpace       ID=0x09EF459F Window=0x9A404470 Pos=138,185 Size=1920,1056 Split=X Selected=0x13926F0B
   DockNode      ID=0x00000003 Parent=0x09EF459F SizeRef=1514,1056 Split=X
     DockNode    ID=0x00000001 Parent=0x00000003 SizeRef=1219,1056 Split=Y Selected=0x13926F0B
       DockNode  ID=0x00000005 Parent=0x00000001 SizeRef=1219,748 CentralNode=1 Selected=0x13926F0B
